requirements.txt
streamlit==1.42.0
pandas==2.2.3
folium==0.19.4
streamlit-folium==0.24.0
altair==5.5.0
numpy==1.26.4
google-genai==0.3.0
duckdb==1.1.3
playwright==1.49.1
selenium==4.27.1
